Revert "Revert "Don't force use of chromedriver-helper""
This reverts commit c42c2ac4214cd03360ac7b7c986fc49d99c8a95a.

This was reverted as installing older versions of the
chromedriver-helper gem caused problems. Unfortunately, the wraith gem
has an unnecessarily strict dependency on the old version of the gem.

I was waiting until wraith was updated to fix this, but I think
service-manual-frontend is unnecessarily installing wraith on the CI
environment, which is triggering this issue. Therefore, just avoiding
unnecessary installs of wraith should be enough to avoid the problems
previously encountered with this.