Bump the version to 0.4.0
This is a repeat of the 0.3.0 release, in which the change was
reverted due to conflicts between older versions of the
chromedriver-helper gem being installed in the CI environment.

I believe the cause of this was the dependency the
service-manual-frontend has on wraith, that has a
strict (unnecessarily so) dependency on the older version of
chromedriver-helper.

This has now been changed, and the older version of
chromedriver-helper isn't appearing on the CI agents anymore.