Support explicitly using the system chromedriver
Looking up `chromedriver` on the PATH is a good idea in principle, but
the reality of how other gems (like wraith) that depend on old
versions of the chromedriver-helper gem which also provide a
non-working version of this binary, combined by the lack of any
isolation between different applications on the GOV.UK CI system makes
this error prone.

Therefore, just allow toggling this behaviour on via an environment
variable.