Remove chromedriver-helper and use system binaries
Previously we used the chromedriver-helper gem to automatically download
a suitable chromedriver binary for use with headless testing. This gem
is no longer maintained, and since work had already begun to start using
the system binary, it seems best to drop it altogether.

https://github.com/flavorjones/chromedriver-helper/issues/83

A system binary for chromedriver should already installed on all Jenkins
agents, but a Puppet gem is still needed to install it on all Dev VMs.