Use Selenium::WebDriver::Chrome::Options
The previous approach we were using to configure Chrome was not
compatible with Chrome 75.0.3770.0.90 and up. This is because the key
had changed from "chromeOptions" to "goog:chromeOptions" in Chromedriver
2.31 and must have recently lost backwards compatibility [1]

The symptoms of this breaking are tests that either don't run at all,
with errors like:

```
Selenium::WebDriver::Error::UnknownError:
  unknown error: Chrome failed to start: exited abnormally
    (unknown error: DevToolsActivePort file doesn't exist)
    (The process started from chrome location /usr/bin/google-chrome is no longer running, so ChromeDriver is assuming that Chrome has crashed.)
```

or tests that run in a non headless version of Chrome (essentially no
arguments passed)

If we had been using the Selenium::WebDriver::Chrome::Options class this
change would have been covered automatically, but since we configured it
manually we didn't realise about this change until it actually broke.

This changes the code to use Selenium::WebDriver::Chrome::Options to
configure the tests. It can no longer accept a hash of options as per
the most recent change, to meet this need this can accept an instance of
`Selenium::WebDriver::Chrome::Options` which can have options
pre-specified.

This removes the window_size option, I'm not sure if any GOV.UK apps
make use of it (I couldn't find any) and this could be configured on a
`Selenium::WebDriver::Chrome::Options` instance.

The `disable-gpu` argument has been removed too. This is no longer
required for mac/linux machines [2], and the bug blocking Windows has
also been resolved for a year [3]. So in the unlikely event that someone
runs this on windows, that is probably good too.

Finally, if we were to make the jump that custom Chrome configuration is
something that should be done outside this gem we could really make this
simpler by removing all the Chrome config and having just
`Capybara.javascript_driver = :selenium_chrome_headless` which is
automatically configured by Capybara.

[1]: https://github.com/SeleniumHQ/docker-selenium/issues/674
[2]: https://github.com/rails/rails/pull/32488
[3]: https://bugs.chromium.org/p/chromium/issues/detail?id=729961