Add Brakeman as a dependency
This will allow us to use the version of Brakeman specified in the
`Gemfile.lock` for the project rather than the version manually
installed onto the CI or our local machines. Currently we can't do this
because most of our apps don't have Brakeman listed as a dependency.

This will make it easier to run Brakeman locally when debugging security
issues that get flagged by the tool, but should also make the CI run
more reliable as all we need to do is run `bundle exec brakeman` which
is easier to understand and less complex than the current method (of
manually installing Brakeman first into a directory and running it from
there).

I'm also hoping this will fix an intermittent problem we have where
sometimes Brakeman will fail to run correctly in CI as it hasn't been
installed properly.