Allow chrome to run outside a sandbox with env var
It is desirable for us to run chrome as the root user in docker
containers, this is because docker tends to favour the root user and it
can be difficult to set up volumes without them being owned by root.
However to run Chrome as the root user the no sandbox option needs to be
provided. This option disables protections that Chrome has to
effectively jail processes that individual web tabs use to stop them
intercommunicating or interacting with the underlying system [1].

To minimise the risk of this affecting users machines inadvertently I've
set this option to only be switched on via an environment variable. This
makes it entirely opt-in. I then expect to set this environment variable
within docker containers so that it is turned on automatically in those
environments.

[1]: https://www.google.com/googlebooks/chrome/med_26.html