Provide a `.headless_chrome_selenium_options` to access options
This method is provided so that these options can be shared with other
utilities, most notably for configuring Jasmine, to avoid duplicating the same
options in two places of an app. This would change that code that looks like:

```
class ChromeHeadlessJasmineConfigurer < JasmineSeleniumRunner::ConfigureJasmine
  def selenium_options
    chrome_options = Selenium::WebDriver::Chrome::Options.new
    chrome_options.headless!
    { options: chrome_options }
  end
end
```

to

```
class ChromeHeadlessJasmineConfigurer < JasmineSeleniumRunner::ConfigureJasmine
  def selenium_options
    { options: GovukTest.headless_chrome_selenium_options }
  end
end
```

This also removes the functionality for apps to pass an argument into
configure that specifies the Selenium::WebDriver::Chrome::Options. This
was removed since configure no longer manipulates options. This feature
was only used by one application [1] that is using a legacy version.

This change addresses a flaw that passing options doesn't cater for all
the configuration that might be needed to specify a Selenium driver,
which can be evidenced by publishing-e2e-tests [2].

In light of this the new recommended process to add custom options is
to redefine the :headless_chrome driver for example:

```
GovukTest.configure

Capybara.register_driver :headless_chrome do |app|
  chrome_options = GovukTest.headless_chrome_selenium_options
  chrome_options.add_argument("--disable-web-security")
  chrome_options.add_argument("--window-size=1400,1400")

  Capybara::Selenium::Driver.new(
    app,
    browser: :chrome,
    desired_capabilities: { acceptInsecureCerts: true },
    options: chrome_options,
  )
end
```

[1]: https://github.com/alphagov/datagovuk_find/blob/76a78a6470ed6953864a432e6f5912b852f0beb9/spec/support/govuk_test.rb#L1
[2]: https://github.com/alphagov/publishing-e2e-tests/blob/43ac18bba872c8f73e5a543190cfc787a86c38bb/spec/spec_helper.rb#L117-L130