Remove Pact as a dependency
We now rely on this being defined in downstream apps which call
the pact:verify:branch task. Apps which don't have Pact installed
are never going to need to execute the rake task, so it is a safe
assumption that Pact should already be available at the time of
calling the task.

The benefit of this approach is that apps which don't have any use
for Pact don't need to download it as a dependency. Prior to this
commit, any app making use of govuk_test would automatically
download the sizeable Pact dependency (which has a dependency on
around 15 gems, including RSpec, which itself is a sizeable gem).